Move test, fix old test
- moved new test to other file to group similar test
- fix old test that was failing due to change
